Move the QuerySet search method into a mixin
