+++ wagtail/wagtaildocs/models.py
@@ -16,18 +16,11 @@ from django.utils.encoding import python_2_unicode_compatible
 from wagtail.wagtailadmin.taggable import TagSearchable
 from wagtail.wagtailadmin.utils import get_object_usage
 from wagtail.wagtailsearch import index
-from wagtail.wagtailsearch.backends import get_search_backend
+from wagtail.wagtailsearch.queryset import SearchableQuerySetMixin
 
 
-class DocumentQuerySet(models.QuerySet):
-    def search(self, query_string, fields=None,
-            operator=None, order_by_relevance=True, backend='default'):
-        """
-        This runs a search query on all the documents in the QuerySet
-        """
-        search_backend = get_search_backend(backend)
-        return search_backend.search(query_string, self, fields=fields,
-            operator=operator, order_by_relevance=order_by_relevance)
+class DocumentQuerySet(SearchableQuerySetMixin, models.QuerySet):
+    pass
 
 
 @python_2_unicode_compatible
