+++ wagtail/wagtailimages/models.py
@@ -28,7 +28,7 @@ from unidecode import unidecode
 from wagtail.wagtailcore import hooks
 from wagtail.wagtailadmin.taggable import TagSearchable
 from wagtail.wagtailsearch import index
-from wagtail.wagtailsearch.backends import get_search_backend
+from wagtail.wagtailsearch.queryset import SearchableQuerySetMixin
 from wagtail.wagtailimages.rect import Rect
 from wagtail.wagtailimages.exceptions import InvalidFilterSpecError
 from wagtail.wagtailadmin.utils import get_object_usage
@@ -42,15 +42,8 @@ class SourceImageIOError(IOError):
     pass
 
 
-class ImageQuerySet(models.QuerySet):
-    def search(self, query_string, fields=None,
-            operator=None, order_by_relevance=True, backend='default'):
-        """
-        This runs a search query on all the images in the QuerySet
-        """
-        search_backend = get_search_backend(backend)
-        return search_backend.search(query_string, self, fields=fields,
-            operator=operator, order_by_relevance=order_by_relevance)
+class ImageQuerySet(SearchableQuerySetMixin, models.QuerySet):
+    pass
 
 
 def get_upload_to(instance, filename):
